chore(packages/feature-flags): rename index to is-feature-active
