+++ packages/feature-flags/index.js
@@ -1,1 +1,1 @@
-export * from './src/index.js';
+export * from './src/is-feature-active.js';
